+++ bee-api/bee-rest-api/src/handlers/tips.rs
@@ -1,0 +1,33 @@
+// Copyright 2020 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+use crate::{
+    filters::CustomRejection::ServiceUnavailable,
+    handlers::{EnvelopeContent, SuccessEnvelope},
+    storage::Backend,
+};
+use bee_common::node::ResHandle;
+use bee_protocol::tangle::MsTangle;
+use serde::Serialize;
+use warp::{reject, Rejection, Reply};
+
+pub(crate) async fn tips<B: Backend>(tangle: ResHandle<MsTangle<B>>) -> Result<impl Reply, Rejection> {
+    match tangle.get_messages_to_approve().await {
+        Some(tips) => Ok(warp::reply::json(&SuccessEnvelope::new(GetTipsResponse {
+            tip_1_message_id: tips.0.to_string(),
+            tip_2_message_id: tips.1.to_string(),
+        }))),
+        None => Err(reject::custom(ServiceUnavailable("tip pool is empty".to_string()))),
+    }
+}
+
+/// Response of GET /api/v1/tips
+#[derive(Clone, Debug, Serialize)]
+pub struct GetTipsResponse {
+    #[serde(rename = "tip1MessageId")]
+    pub tip_1_message_id: String,
+    #[serde(rename = "tip2MessageId")]
+    pub tip_2_message_id: String,
+}
+
+impl EnvelopeContent for GetTipsResponse {}
